FEATURE: Separate NodeCreation into creating "commands" (NodeMutators)
We queue changes on nodes instead of making them directly.
This allows us to check if the template is valid by cr perspective, without actually applying it,
as all checks are executed before the Mutators are run.
A mutator should only queue a mutation, that is guaranteed to succeed, as there is no exception handing.